// src/enrollments/enums/enrollment.enums.ts
export enum EnrollmentSource {
  FREE = 'FREE',
  PAID = 'PAID',
  SEAT = 'SEAT',
}

export enum EnrollmentStatus {
  ACTIVE = 'ACTIVE',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
}

export enum LessonProgressStatus {
  NOT_STARTED = 'NOT_STARTED',
  IN_PROGRESS = 'IN_PROGRESS',
  COMPLETED = 'COMPLETED',
}